Conflicts:
	features/topology-map/plugins/org.opennms.features.topology.plugins.topo.linkd/src/main/java/org/opennms/features/topology/plugins/topo/linkd/internal/EnhancedLinkdTopologyProvider.java
	features/topology-map/plugins/org.opennms.features.topology.plugins.topo.linkd/src/main/resources/OSGI-INF/blueprint/blueprint.xml
	features/topology-map/plugins/org.opennms.features.topology.plugins.topo.linkd/src/test/java/org/opennms/features/topology/plugins/topo/linkd/internal/EnhancedLinkdMockDataPopulator.java
	features/topology-map/plugins/org.opennms.features.topology.plugins.topo.linkd/src/test/resources/META-INF/opennms/applicationContext-enhanced-mock.xml
	opennms-dao-api/src/main/java/org/opennms/netmgt/dao/api/BridgeTopologyDao.java
	opennms-dao/src/main/java/org/opennms/netmgt/dao/hibernate/BridgeMacLinkDaoHibernate.java
	opennms-dao/src/main/java/org/opennms/netmgt/dao/hibernate/BridgeTopologyDaoHibernate.java
	opennms-dao/src/main/java/org/opennms/netmgt/dao/hibernate/CdpLinkDaoHibernate.java
	opennms-dao/src/main/java/org/opennms/netmgt/dao/hibernate/IsIsLinkDaoHibernate.java
	opennms-model/src/main/java/org/opennms/netmgt/model/BridgeBridgeLink.java
	opennms-model/src/main/java/org/opennms/netmgt/model/BridgeMacLink.java
	opennms-model/src/main/java/org/opennms/netmgt/model/topology/Bridge.java
	opennms-model/src/main/java/org/opennms/netmgt/model/topology/BridgePort.java
	opennms-model/src/main/java/org/opennms/netmgt/model/topology/BroadcastDomain.java
	opennms-model/src/main/java/org/opennms/netmgt/model/topology/SharedSegment.java
	opennms-services/src/main/java/org/opennms/netmgt/enlinkd/EnhancedLinkd.java
	opennms-services/src/main/java/org/opennms/netmgt/enlinkd/EnhancedLinkdService.java
	opennms-services/src/main/java/org/opennms/netmgt/enlinkd/EnhancedLinkdServiceImpl.java
	opennms-services/src/main/java/org/opennms/netmgt/enlinkd/NodeDiscoveryBridge.java
	opennms-services/src/main/java/org/opennms/netmgt/enlinkd/NodeDiscoveryBridgeTopology.java
	opennms-services/src/main/java/org/opennms/netmgt/enlinkd/snmp/Dot1qTpFdbTableTracker.java
	opennms-services/src/test/java/org/opennms/netmgt/enlinkd/EnLinkdIT.java
	opennms-services/src/test/java/org/opennms/netmgt/enlinkd/EnLinkdSnmpIT.java
	opennms-services/src/test/java/org/opennms/netmgt/enlinkd/EnLinkdTestHelper.java
	opennms-webapp/src/main/java/org/opennms/web/enlinkd/NodeLinkBridge.java

+++ opennms-model/src/main/java/org/opennms/netmgt/model/OnmsAssetRecord.java
@@ -1654,12 +1654,12 @@ public class OnmsAssetRecord implements Serializable {
     @Transient
     @Deprecated
     @XmlElement
-    public Float getLongitude() {
+    public Double getLongitude() {
         return m_geolocation == null ? null : m_geolocation.getLongitude();
     }
 
     @Deprecated
-    public void setLongitude(final Float longitude) {
+    public void setLongitude(final Double longitude) {
         if (m_geolocation != null)
             m_geolocation.setLongitude(longitude);
     }
@@ -1670,12 +1670,12 @@ public class OnmsAssetRecord implements Serializable {
     @Transient
     @Deprecated
     @XmlElement
-    public Float getLatitude() {
+    public Double getLatitude() {
         return m_geolocation == null ? null : m_geolocation.getLatitude();
     }
 
     @Deprecated
-    public void setLatitude(final Float latitude) {
+    public void setLatitude(final Double latitude) {
         if (m_geolocation != null)
             m_geolocation.setLatitude(latitude);
     }
